+++ repository/src/main/java/com/intive/repository/Repository.kt
@@ -24,6 +24,7 @@ interface Repository {
     suspend fun getUsers(page: Int, role: String, group: String?): UsersResponse
     suspend fun getUsers(page: Int, role: String, group: String?, firstName: String?, lastName: String?): UsersResponse
     suspend fun getUsers(page: Int, role: String, group: String?, firstName: String?, lastName: String?, login: String?): UsersResponse
+    suspend fun getUsers(page: Int, role: String, group: String?, query: String): UsersResponse
     suspend fun getTotalUsersByRole(role: String, group: String?): Int
 
     suspend fun searchAudits(page: Int, query: String): List<Audit>
