+++ repository/src/main/java/com/intive/repository/RepositoryImpl.kt
@@ -10,6 +10,7 @@ import com.intive.repository.network.util.EventDtoMapper
 
 import com.intive.repository.domain.model.Audit
 import com.intive.repository.domain.model.EventInviteResponse
+import com.intive.repository.network.ROLE_CANDIDATE
 import com.intive.repository.network.util.AuditDtoMapper
 import com.intive.repository.network.util.EventInviteResponseDtoMapper
 import com.intive.repository.network.response.UsersResponse
@@ -72,6 +73,44 @@ class RepositoryImpl(
             lastName = lastName,
             login = login
         )
+    }
+
+    override suspend fun getUsers(
+        page: Int,
+        role: String,
+        group: String?,
+        query: String,
+    ): UsersResponse {
+
+        return when {
+            query.isBlank() -> {
+                networkRepository.getUsers(
+                    page = page,
+                    role = role,
+                    group = group
+                )
+            }
+            query.split(" ").size == 1 -> {
+                networkRepository.getUsers(
+                    page = page,
+                    role = role,
+                    group = group,
+                    firstName = query,
+                    lastName = query,
+                    login = query
+                )
+            }
+            else -> {
+                val q = query.split(" ")
+                networkRepository.getUsers(
+                    page = page,
+                    role = role,
+                    group = group,
+                    firstName = q[0],
+                    lastName = q[1]
+                )
+            }
+        }
     }
 
     override suspend fun getTotalUsersByRole(role: String, group: String?): Int {
